mainly renaming variables and some simplyfing
